MUN-40 refactor:코드 리팩토링 중
