+++ src/main/java/ureca/muneobe/common/addon/dto/request/DefaultAddonCreateRequest.java
@@ -5,7 +5,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import ureca.muneobe.common.chat.entity.AddonType;
+import ureca.muneobe.common.addon.entity.AddonType;
 
 @Getter
 @Builder
