+++ src/main/java/ureca/muneobe/common/addon/dto/request/DefaultAddonSearchRequest.java
@@ -1,0 +1,14 @@
+package ureca.muneobe.common.addon.dto.request;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class DefaultAddonSearchRequest {
+    private String name;
+}
