+++ src/main/java/ureca/muneobe/common/addon/dto/response/DefaultAddonResponse.java
@@ -6,7 +6,7 @@ import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import ureca.muneobe.common.addon.entity.DefaultAddon;
-import ureca.muneobe.common.chat.entity.AddonType;
+import ureca.muneobe.common.addon.entity.AddonType;
 
 @Getter
 @Builder
