+++ src/main/java/ureca/muneobe/common/addon/repository/DefaultAddonRepository.java
@@ -1,5 +1,6 @@
 package ureca.muneobe.common.addon.repository;
 
+import java.util.Optional;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import ureca.muneobe.common.addon.entity.DefaultAddon;
@@ -7,4 +8,5 @@ import ureca.muneobe.common.addon.entity.DefaultAddon;
 
 @Repository
 public interface DefaultAddonRepository extends JpaRepository<DefaultAddon, Long> {
+    Optional<DefaultAddon> findByName(String name);
 }
