+++ src/main/java/ureca/muneobe/common/addongroup/repository/AddonGroupRepository.java
@@ -1,7 +1,7 @@
 package ureca.muneobe.common.addongroup.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import ureca.muneobe.common.chat.entity.AddonGroup;
+import ureca.muneobe.common.addongroup.entity.AddonGroup;
 
 public interface AddonGroupRepository extends JpaRepository<AddonGroup, Long> {
 }
