--- src/main/java/ureca/muneobe/common/chat/repository/MplanRepository.java
@@ -1,7 +1,0 @@
-package ureca.muneobe.common.chat.repository;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-import ureca.muneobe.common.chat.entity.Mplan;
-
-public interface MplanRepository extends JpaRepository<Mplan, Long> {
-}
