+++ src/main/java/ureca/muneobe/common/chat/repository/search/AddonSearchRepository.java
@@ -5,10 +5,9 @@ import com.querydsl.core.Tuple;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Repository;
-import ureca.muneobe.common.chat.entity.AddonType;
-import ureca.muneobe.common.chat.entity.QAddon;
-import ureca.muneobe.common.chat.entity.QAddonGroup;
-import ureca.muneobe.common.chat.entity.QMplan;
+import ureca.muneobe.common.addon.entity.AddonType;
+import ureca.muneobe.common.addon.entity.QAddon;
+import ureca.muneobe.common.addongroup.entity.QAddonGroup;
 import ureca.muneobe.common.chat.service.strategy.rdb.input.AddonCondition;
 import ureca.muneobe.common.chat.service.strategy.rdb.input.Condition;
 import ureca.muneobe.common.chat.service.strategy.rdb.input.Range;
@@ -17,6 +16,7 @@ import ureca.muneobe.common.chat.service.strategy.rdb.output.FindingMplan;
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
+import ureca.muneobe.common.mplan.entity.QMplan;
 
 import static ureca.muneobe.common.chat.repository.search.SearchUtils.*;
 
@@ -69,7 +69,7 @@ public class AddonSearchRepository implements SearchRepository {
         List<Long> addonGroupIds = jpaQueryFactory
                 .select(addonGroup.id)
                 .from(addonGroup)
-                .join(addonGroup.addon, addon)
+                .join(addonGroup.addons, addon)
                 .where(addon.name.in(names))
                 .groupBy(addonGroup.id)
                 .having(addon.name.countDistinct().eq(requiredCount))
@@ -102,7 +102,7 @@ public class AddonSearchRepository implements SearchRepository {
             if (priceRange != null) {
                 applyRange(on, a.price, priceRange);
             }
-            query = query.leftJoin(addonGroup.addon, a).on(on);
+            query = query.leftJoin(addonGroup.addons, a).on(on);
         }
 
         // 모두 매칭되었음을 확인
@@ -134,7 +134,7 @@ public class AddonSearchRepository implements SearchRepository {
         List<Long> addonGroupIds = jpaQueryFactory
                 .select(addonGroup.id)
                 .from(addonGroup)
-                .join(addonGroup.addon, addon)
+                .join(addonGroup.addons, addon)
                 .where(predicate)
                 .distinct()
                 .fetch();
