+++ src/main/java/ureca/muneobe/common/chat/service/strategy/rdb/output/FindingAddon.java
@@ -4,7 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.ToString;
-import ureca.muneobe.common.chat.entity.AddonType;
+import ureca.muneobe.common.addon.entity.AddonType;
 
 @AllArgsConstructor
 @Builder
