+++ src/main/java/ureca/muneobe/common/combined/Combined.java
@@ -1,4 +1,4 @@
-package ureca.muneobe.common.chat.entity;
+package ureca.muneobe.common.combined;
 
 import jakarta.persistence.*;
