+++ src/main/java/ureca/muneobe/common/mplan/dto/response/AddonGroupAddonResponse.java
@@ -5,8 +5,8 @@ import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
-import ureca.muneobe.common.chat.entity.Addon;
-import ureca.muneobe.common.chat.entity.AddonType;
+import ureca.muneobe.common.addon.entity.Addon;
+import ureca.muneobe.common.addon.entity.AddonType;
 
 @Getter
 @Builder
