+++ src/main/java/ureca/muneobe/common/mplan/dto/response/MplanDetailsResponse.java
@@ -2,7 +2,7 @@ package ureca.muneobe.common.mplan.dto.response;
 
 import lombok.*;
 import org.springframework.data.domain.Page;
-import ureca.muneobe.common.chat.entity.MplanDetail;
+import ureca.muneobe.common.mplan.entity.MplanDetail;
 
 @Getter
 @Builder
