+++ src/main/java/ureca/muneobe/common/mplan/entity/MplanType.java
@@ -1,4 +1,4 @@
-package ureca.muneobe.common.chat.entity;
+package ureca.muneobe.common.mplan.entity;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
