+++ src/main/java/ureca/muneobe/common/mplan/repository/MplanDetailRepository.java
@@ -1,7 +1,7 @@
 package ureca.muneobe.common.mplan.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import ureca.muneobe.common.chat.entity.MplanDetail;
+import ureca.muneobe.common.mplan.entity.MplanDetail;
 
 public interface MplanDetailRepository extends JpaRepository<MplanDetail, Long> {
 }
