+++ src/main/java/ureca/muneobe/common/mplan/repository/MplanRepository.java
@@ -1,7 +1,7 @@
 package ureca.muneobe.common.mplan.repository;
 
 import org.springframework.data.jpa.repository.JpaRepository;
-import ureca.muneobe.common.chat.entity.Mplan;
+import ureca.muneobe.common.mplan.entity.Mplan;
 
 public interface MplanRepository extends JpaRepository<Mplan, Long> {
 }
