+++ src/main/java/ureca/muneobe/common/mplan/service/MplanService.java
@@ -7,9 +7,9 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import ureca.muneobe.common.addongroup.repository.AddonGroupRepository;
 import ureca.muneobe.common.auth.entity.enums.Role;
-import ureca.muneobe.common.chat.entity.AddonGroup;
-import ureca.muneobe.common.chat.entity.Mplan;
-import ureca.muneobe.common.chat.entity.MplanDetail;
+import ureca.muneobe.common.addongroup.entity.AddonGroup;
+import ureca.muneobe.common.mplan.entity.Mplan;
+import ureca.muneobe.common.mplan.entity.MplanDetail;
 import ureca.muneobe.common.mplan.dto.request.MplanCreateRequest;
 import ureca.muneobe.common.mplan.dto.response.MplanCreateResponse;
 import ureca.muneobe.common.mplan.dto.response.MplansResponse;
