+++ src/main/java/ureca/muneobe/common/mypage/controller/MyPageController.java
@@ -1,12 +1,11 @@
 //package ureca.muneobe.common.mypage.controller;
 //
+//import jakarta.servlet.http.HttpSession;
 //import lombok.RequiredArgsConstructor;
 //import org.springframework.http.ResponseEntity;
 //import org.springframework.web.bind.annotation.GetMapping;
-//import org.springframework.web.bind.annotation.RequestParam;
 //import org.springframework.web.bind.annotation.RestController;
-//import ureca.muneobe.common.mypage.dto.MyPageResponse;
-//import ureca.muneobe.common.mypage.service.MyPageService;
+//import ureca.muneobe.common.auth.utils.SessionUtil;
 //import ureca.muneobe.global.response.ResponseBody;
 //
 //@RestController
@@ -16,8 +15,9 @@
 //
 //    @GetMapping("/mypage")
 //    public ResponseEntity<ResponseBody<MyPageResponse>> readMyPage(
-//            @RequestParam(defaultValue = "0") int page
+//            HttpSession httpSession
 //    ){
-//
+//        return ResponseEntity.ok().body(
+//                ResponseBody.success(myPageService.findMyInfo(SessionUtil.getLoginMember(httpSession))));
 //    }
 //}
