--- src/main/java/ureca/muneobe/common/mypage/dto/MyPageResponse.java
@@ -1,4 +1,0 @@
-package ureca.muneobe.common.mypage.dto;
-
-public class MyPageResponse {
-}
