+++ src/main/java/ureca/muneobe/common/mypage/dto/response/MyPageResponse.java
@@ -1,0 +1,19 @@
+//package ureca.muneobe.common.mypage.dto.response;
+//
+//import java.time.LocalDateTime;
+//import lombok.AllArgsConstructor;
+//import lombok.Builder;
+//import lombok.Getter;
+//import lombok.NoArgsConstructor;
+//
+//@Getter
+//@Builder
+//@NoArgsConstructor
+//@AllArgsConstructor
+//public class MyPageResponse {
+//    private String name;
+//    private
+//    private String phoneNumber;
+//    private LocalDateTime createdAt;
+//    private
+//}
