+++ src/main/java/ureca/muneobe/common/mypage/service/MyPageService.java
@@ -1,15 +1,19 @@
-package ureca.muneobe.common.mypage.service;
-
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Service;
-import ureca.muneobe.common.auth.respository.MemberRepository;
-import ureca.muneobe.common.chat.entity.Subscription;
-
-@Service
-@RequiredArgsConstructor
-public class MyPageService {
-    private final MemberRepository memberRepository;
-    private final SubscriptionRepository subscriptionRepository;
-
-
-}
+//package ureca.muneobe.common.mypage.service;
+//
+//import lombok.RequiredArgsConstructor;
+//import org.springframework.stereotype.Service;
+//import ureca.muneobe.common.auth.entity.Member;
+//import ureca.muneobe.common.auth.respository.MemberRepository;
+//import ureca.muneobe.common.mypage.dto.response.MyPageResponse;
+//
+//@Service
+//@RequiredArgsConstructor
+//public class MyPageService {
+//    private final MemberRepository memberRepository;
+//    private final SubscriptionRepository subscriptionRepository;
+//
+//
+//    public MyPageResponse findMyInfo(Member loginMember) {
+//        return null;
+//    }
+//}
