+++ src/main/java/ureca/muneobe/common/review/entity/Review.java
@@ -1,4 +1,5 @@
-package ureca.muneobe.common.chat.entity;
+package ureca.muneobe.common.review.entity;
+
 
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
@@ -13,28 +14,25 @@ import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import ureca.muneobe.common.auth.entity.Member;
+import ureca.muneobe.common.mplan.entity.Mplan;
 import ureca.muneobe.global.common.BaseEntity;
 
 @Entity
-@Table(name = "subscription")
+@Table(name = "Review")
 @Getter
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Subscription extends BaseEntity {
+public class Review extends BaseEntity {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
     private Long id;
 
-    @Column(name = "fee")
-    private Integer fee;
+    @Column(name = "content")
+    private String content;
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "member_id")
+    @JoinColumn(name = "user_id")
     private Member member;
-
-    @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "combined_group_id")
-    private CombinedGroup combinedGroup;
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "mplan_id")
