+++ src/test/java/ureca/muneobe/common/chat/repository/search/CombinedSearchRepositoryTest.java
@@ -10,10 +10,10 @@
 //import org.testcontainers.containers.PostgreSQLContainer;
 //import org.testcontainers.junit.jupiter.Container;
 //import org.testcontainers.junit.jupiter.Testcontainers;
-//import ureca.muneobe.common.chat.entity.AddonType;
-//import ureca.muneobe.common.chat.entity.DataType;
-//import ureca.muneobe.common.chat.entity.MplanType;
-//import ureca.muneobe.common.chat.entity.Qualification;
+//import ureca.muneobe.common.addon.entity.AddonType;
+//import ureca.muneobe.common.mplan.entity.DataType;
+//import ureca.muneobe.common.mplan.entity.MplanType;
+//import ureca.muneobe.common.mplan.entity.Qualification;
 //import ureca.muneobe.common.chat.service.rdb.input.AddonCondition;
 //import ureca.muneobe.common.chat.service.rdb.input.Condition;
 //import ureca.muneobe.common.chat.service.rdb.input.MplanCondition;
